+++ src/Faker/Internal/Types.hs
@@ -27,6 +27,7 @@ data SourceData
   | Chiquito
   | Computer
   | Control
+  | DnD
   | Departed
   | BossaNova
   | BreakingBad