+++ src/Faker/Provider/Dnd.hs
@@ -1,0 +1,76 @@
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TemplateHaskell #-}
+
+module Faker.Provider.Dnd where
+
+import Config
+import Control.Monad.Catch
+import Data.Text (Text)
+import Data.Vector (Vector)
+import Data.Monoid ((<>))
+import Data.Yaml
+import Faker
+import Faker.Internal
+import Faker.Provider.TH
+import Language.Haskell.TH
+
+parseDnd :: FromJSON a => FakerSettings -> Value -> Parser a
+parseDnd settings (Object obj) = do
+  en <- obj .: (getLocale settings)
+  faker <- en .: "faker"
+  dnd <- faker .: "dnd"
+  pure dnd
+parseDnd settings val = fail $ "expected Object, but got " <> (show val)
+
+parseDndField ::
+     (FromJSON a, Monoid a) => FakerSettings -> Text -> Value -> Parser a
+parseDndField settings txt val = do
+  dnd <- parseDnd settings val
+  field <- dnd .:? txt .!= mempty
+  pure field
+
+parseDndFields ::
+     (FromJSON a, Monoid a) => FakerSettings -> [Text] -> Value -> Parser a
+parseDndFields settings txts val = do
+  dnd <- parseDnd settings val
+  helper dnd txts
+  where
+    helper :: (FromJSON a) => Value -> [Text] -> Parser a
+    helper a [] = parseJSON a
+    helper (Object a) (x:xs) = do
+      field <- a .: x
+      helper field xs
+    helper a (x:xs) = fail $ "expect Object, but got " <> (show a)
+
+
+
+
+$(genParser "dnd" "species")
+
+$(genProvider "dnd" "species")
+
+
+$(genParser "dnd" "klasses")
+
+$(genProvider "dnd" "klasses")
+
+
+$(genParser "dnd" "backgrounds")
+
+$(genProvider "dnd" "backgrounds")
+
+
+$(genParser "dnd" "alignments")
+
+$(genProvider "dnd" "alignments")
+
+
+
+
+
+
+
+
+
+
+